+++ src/WiFiManager_NINA_Lite_SAMD.h
@@ -8,31 +8,16 @@
 
   Built by Khoi Hoang https://github.com/khoih-prog/WiFiManager_NINA_Lite
   Licensed under MIT license
-  Version: 1.6.2
+  Version: 1.6.3
 
   Version Modified By   Date        Comments
   ------- -----------  ----------   -----------
   1.0.0   K Hoang      26/03/2020  Initial coding
-  1.0.1   K Hoang      27/03/2020  Fix SAMD soft-reset bug. Add support to remaining boards
-  1.0.2   K Hoang      15/04/2020  Fix bug. Add SAMD51 support.
-  1.0.3   K Hoang      24/04/2020  Fix bug. Add nRF5 (Adafruit, NINA_B302_ublox, etc.) support. Add MultiWiFi, HostName capability.
-                                   SSID password maxlen is 63 now. Permit special chars # and % in input data.
-  1.0.4   K Hoang      04/05/2020  Add Configurable Config Portal Title, Default Config Data and DRD. Update examples.
-  1.0.5   K Hoang      11/07/2020  Modify LOAD_DEFAULT_CONFIG_DATA logic. Enhance MultiWiFi connection logic. Add MQTT examples.
-  1.1.0   K Hoang      19/02/2021  Optimize code and use better FlashStorage_SAMD and FlashStorage_STM32.
-                                   Add customs HTML header feature. Fix bug.
-  1.1.1   K Hoang      13/03/2021  Fix USE_DYNAMIC_PARAMETERS bug.
-  1.1.2   K Hoang      30/03/2021  Fix MultiWiFi connection bug.
-  1.1.3   K Hoang      12/04/2021  Fix invalid "blank" Config Data treated as Valid.
-  1.2.0   K Hoang      14/04/2021  Optional one set of WiFi Credentials. Enforce WiFi PWD minimum 8 chars
-  1.3.0   Michael H    21/04/2021  Enable scan of WiFi networks for selection in Configuration Portal
-  1.3.1   K Hoang      15/05/2021  Fix createHTML bug.
-  1.4.0   K Hoang      28/05/2021  Add support to Nano_RP2040_Connect, RASPBERRY_PI_PICO using Arduino mbed or Arduino-pico core
-  1.4.1   K Hoang      12/10/2021  Update `platform.ini` and `library.json`
-  1.5.0   K Hoang      05/01/2022  Fix the blocking issue in loop()
+  ...
   1.6.0   K Hoang      05/01/2022  Configurable WIFI_RECON_INTERVAL. Add support to RP2040 using arduino-pico core
   1.6.1   K Hoang      26/01/2022  Update to be compatible with new FlashStorage libraries. Add support to more SAMD/STM32 boards
   1.6.2   K Hoang      18/02/2022  Optional Board_Name in Menu. Optimize code by using passing by reference
+  1.6.3   K Hoang      20/02/2022  Add optional CONFIG_MODE_LED. Add function to
   **********************************************************************************************************************************/
 
 #ifndef WiFiManager_NINA_Lite_SAMD_h
@@ -55,13 +40,13 @@
 #endif
 
 #ifndef WIFIMANAGER_NINA_LITE_VERSION
-  #define WIFIMANAGER_NINA_LITE_VERSION            "WiFiManager_NINA_Lite v1.6.2"
+  #define WIFIMANAGER_NINA_LITE_VERSION            "WiFiManager_NINA_Lite v1.6.3"
 
   #define WIFIMANAGER_NINA_LITE_VERSION_MAJOR      1
   #define WIFIMANAGER_NINA_LITE_VERSION_MINOR      6
-  #define WIFIMANAGER_NINA_LITE_VERSION_PATCH      2
+  #define WIFIMANAGER_NINA_LITE_VERSION_PATCH      3
 
-#define WIFIMANAGER_NINA_LITE_VERSION_INT        1006002
+#define WIFIMANAGER_NINA_LITE_VERSION_INT          1006003
 
 #endif
 
@@ -360,6 +345,12 @@ class WiFiManager_NINA_Lite
     void begin(const char *iHostname = "")
     {
       #define RETRY_TIMES_CONNECT_WIFI			3
+
+#if USING_CONFIG_MODE_LED  
+      //Turn OFF
+      pinMode(LED_BUILTIN, OUTPUT);
+      digitalWrite(LED_BUILTIN, LED_OFF);
+#endif
       
       if (iHostname[0] == 0)
       {
@@ -515,6 +506,11 @@ class WiFiManager_NINA_Lite
         if (WiFi.status() == WL_CONNECTED)
         {
           wifi_connected = true;
+          
+#if USING_CONFIG_MODE_LED
+          // turn the LED_BUILTIN OFF to tell us we exit configuration mode.
+          digitalWrite(CONFIG_MODE_LED, LED_OFF);
+#endif
         }
         else
         {
@@ -603,6 +599,11 @@ class WiFiManager_NINA_Lite
       {
         configuration_mode = false;
         WN_LOGERROR(F("r:gotWBack"));
+        
+#if USING_CONFIG_MODE_LED
+          // turn the LED_BUILTIN OFF to tell us we exit configuration mode.
+          digitalWrite(CONFIG_MODE_LED, LED_OFF);
+#endif
       }
     }
     
@@ -738,6 +739,13 @@ class WiFiManager_NINA_Lite
     bool isConfigDataValid()
     {
       return hadConfigData;
+    }
+    
+    //////////////////////////////////////////////
+    
+    bool isConfigMode()
+    {
+      return configuration_mode;
     }
     
     //////////////////////////////////////////////
@@ -1943,6 +1951,11 @@ class WiFiManager_NINA_Lite
 	    WiFiNetworksFound = scanWifiNetworks(&indices);	
 #endif
 
+#if USING_CONFIG_MODE_LED
+      // turn the LED_BUILTIN ON to tell us we enter configuration mode.
+      digitalWrite(CONFIG_MODE_LED, LED_ON);
+#endif
+    
       WiFi.config(portal_apIP);
 
       if ( (portal_ssid == "") || portal_pass == "" )